refactor and generic mp
