+++ Manifest.toml
@@ -384,6 +384,11 @@ git-tree-sha1 = "e032c9df551fb23c9f98ae1064de074111b7bc39"
 uuid = "992d4aef-0814-514b-bc4d-f2e9a6c4116f"
 version = "0.3.1"
 
+[[SlidingDistancesBase]]
+path = "/home/fredrikb/.julia/dev/SlidingDistancesBase"
+uuid = "25b0cc0c-38e4-462f-a11d-8564868c562d"
+version = "0.1.0"
+
 [[Sockets]]
 uuid = "6462fe0b-24de-5631-8697-dd941f90decc"
