+++ src/MatrixProfile.jl
@@ -5,6 +5,9 @@ using LoopVectorization
 using ProgressMeter
 using RecipesBase
 
+using SlidingDistancesBase
+import SlidingDistancesBase: floattype, lastlength, distance_profile, distance_profile!
+
 export matrix_profile, distance_profile, motifs
 
 
@@ -22,7 +25,7 @@ Return the matrix profile and the profile indices of time series `T` with window
 
 Reference: [Matrix profile II](https://www.cs.ucr.edu/~eamonn/STOMP_GPU_final_submission_camera_ready.pdf).
 """
-function matrix_profile(T, m; showprogress=true)
+function matrix_profile(T::AbstractVector{<:Number}, m::Int; showprogress=true)
     n   = length(T)
     l   = n-m+1
     n > 2m+1 || throw(ArgumentError("Window length too long, maximum length is $(2m)"))
@@ -35,7 +38,7 @@ function matrix_profile(T, m; showprogress=true)
     prog = Progress((l - 1) ÷ 10, dt=1, desc="Matrix profile", barglyphs = BarGlyphs("[=> ]"), color=:blue)
     @inbounds for i = 2:l
         for j = l:-1:2
-            @fastmath QT[j] = QT[j-1]-T[j-1]*T[i-1]+T[j+m-1]*T[i+m-1]
+            QT[j] = QT[j-1] - T[j-1] * T[i-1] + T[j+m-1] * T[i+m-1]
         end
         QT[1] = QT₀[i]
         distance_profile!(D, QT, μ, σ, m, i)
@@ -50,7 +53,7 @@ end
 
 Compute the Euclidean distance profile corresponding to sliding `Q` over `T`
 """
-function distance_profile!(D, Q, T)
+function distance_profile!(D::AbstractVector{S}, Q::AbstractVector{S}, T::AbstractVector{S}) where S <: Number
     m = length(Q)
     μ,σ  = running_mean_std(T, m)
     QT   = window_dot(znorm(Q), T)
@@ -62,7 +65,7 @@ function distance_profile!(D, Q, T)
 end
 distance_profile(Q, T) = distance_profile!(similar(T, length(T)-length(Q)+1), Q, T)
 
-function distance_profile!(D, QT, μ, σ, m, i)
+function distance_profile!(D::AbstractVector{S}, QT::AbstractVector{S}, μ, σ, m::Int, i::Int) where S <: Number
     @assert i <= length(D)
     @avx for j = eachindex(D)
         frac = (QT[j] - m*μ[i]*μ[j]) / (m*σ[i]*σ[j])
@@ -72,7 +75,7 @@ function distance_profile!(D, QT, μ, σ, m, i)
     D
 end
 
-distance_profile(QT, μ, σ, m) = distance_profile!(similar(QT), QT, μ, σ, m, 1)
+distance_profile(QT::AbstractVector{S}, μ::AbstractVector{S}, σ::AbstractVector{S}, m::Int) where S<:Number = distance_profile!(similar(QT), QT, μ, σ, m, 1)
 
 
 function update_min!(P, I, D, i)
@@ -119,86 +122,38 @@ function running_mean_std(x::AbstractArray{T}, m) where T
     μ,σ
 end
 
-function remove_found_motifs!(P, found_motifs, th)
-    for mo in found_motifs
-        for o in mo.onsets
-            inds = max(1, o-th):min(length(P), o+th)
-            P[inds] .= typemax(eltype(P))
-        end
-    end
-end
-
-function motifs(p::Profile, k, r, th = 0, found_motifs = Motif[])
-    length(found_motifs) == k && return found_motifs
-    m = p.m
-    P = copy(p.P)
-    remove_found_motifs!(P, found_motifs, th)
-    d, i = findmin(P)
-    inds = 0:m-1
-    distance_profile!(P, view(p.T, inds .+ i), p.T)
-    remove_found_motifs!(P, found_motifs, th)
-    onsets = findall(<=(d * r), P)
-    push!(found_motifs, Motif(p, onsets))
-    motifs(p::Profile, k, r, th, found_motifs)
-end
-
-struct Motif
-    motifs
-    onsets
-end
-
-function Motif(p::Profile, onsets)
-    m = p.m
-    inds = 0:m-1
-    n = length(onsets)
-    vecs = map(1:n) do i
-        p.T[inds .+ onsets[i]]
-    end
-    Motif(vecs, onsets)
-end
-
-
-function znorm(x)
+function znorm(x::AbstractVector)
     x = x .- mean(x)
     x ./= std(x, mean=0, corrected=false)
 end
 
-@recipe function plot(p::Profile)
-    link --> :x
-    layout --> (2,1)
-    legend --> false
-    @series begin
-        title --> "T"
-        label --> "T"
-        subplot --> 1
-        p.T
+
+
+## General data and distance ===================================================
+
+
+
+function matrix_profile(T, m::Int, dist; showprogress=true)
+    n   = length(T)
+    l   = n-m+1
+    n > 2m+1 || throw(ArgumentError("Window length too long, maximum length is $(2m)"))
+    P    = distance_profile(dist, getwindow(T,m,1), T)
+    P[1] = typemax(eltype(P))
+    D    = similar(P)
+    I    = ones(Int, l)
+    prog = Progress((l - 1) ÷ 10, dt=1, desc="Matrix profile", barglyphs = BarGlyphs("[=> ]"), color=:blue)
+    @inbounds for i = 2:l
+        Ti = getwindow(T,m,i)
+        distance_profile!(D, dist, Ti, T)
+        update_min!(P, I, D, i)
+        showprogress && i % 10 == 0 && next!(prog)
     end
-    @series begin
-        title --> "Matrix profile"
-        subplot --> 2
-        p.P
-    end
+    Profile(T, P, I, m)
 end
 
-_append_inf(x) = vec([x; fill(Inf, 1, size(x,2))])
 
-@recipe function plot(p::Profile, motifs::Vector{Motif})
+include("motifs.jl")
+include("plotting.jl")
 
-    @series begin
-        p
-    end
-
-    inds = 0:p.m-1
-    linewidth --> 2
-    for (j,m) in enumerate(motifs)
-        @series begin
-            legend --> true
-            group := j
-            label --> j
-            subplot := 1
-            _append_inf(m.onsets' .+ inds), _append_inf(reduce(hcat, m.motifs))
-        end
-    end
-end
 
 end
